+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- fix: incorrect Binance spot order rules calculation [DV-3598]
+- fix: incorrect behaviour of logging [DV-3593]
+- fix: encode analytics data into byte array before forwarding to KV cache [DV-3599]
 
 ## [0.9.9] - 2025-10-14
 - fix: panic in Binance GetOrderRule when handling USDT pairs [DV-3558]